Moves database files into src and updates imports
Unifies database file structure under the source directory for better organization and maintainability.
Updates import paths throughout the codebase to reflect the new locations.
Ensures build and runtime processes use the compiled database files correctly.
