+++ pong-backend/Dockerfile.production
@@ -16,15 +16,12 @@ WORKDIR /app
 
 COPY --from=builder /app/node_modules ./node_modules
 COPY --from=builder /app/package.json ./package.json
-COPY --from=builder /app/dist ./dist
-COPY --from=builder /app/database ./database
 
-# Copy compiled output and any runtime assets
 COPY --from=builder /app/dist ./dist
 COPY --from=builder /app/database ./database
 
 # Initialize the database
-RUN node /app/database/init.js
+RUN node /app/dist/database/init.js
 
 ENV NODE_ENV=production
