+++ pong-backend/src/controllers/profileController.ts
@@ -3,7 +3,7 @@ import { getIdUser, updatedUserInDB } from '../user_service/user_service.js';
 import bcrypt from 'bcrypt';
 import { UpdateBody } from '../types/ProfilType.js';
 import { UsersModel } from '../models/usersModel.js';
-import db from '../../database/db.js'
+import db from '../database/db.js'
 import { print } from '../server.js';
 import { statusCode } from '../types/statusCode.js';
