+++ pong-backend/src/services/chatBlockService.ts
@@ -1,4 +1,4 @@
-import db from "../../database/db.js";
+import db from "../database/db.js";
 import { ChatBlockModel } from "../models/chatBlockModel.js";
 import { print } from "../server.js";
 import { connectedRoomInstance } from "../state/ConnectedRoom.js";
