+++ pong-backend/vitest.setup.ts
@@ -1,3 +1,3 @@
 import { beforeAll, afterAll, describe, it, expect, vi, beforeEach } from 'vitest'
-import { createSchema } from './database/schema';
-import { seedUsers } from './database/seeds/seed_users';
+import { createSchema } from './src/database/schema';
+import { seedUsers } from './src/database/seeds/seed_users';
